create util/types.py to handle types and type checking
